Flags: Drop USED_DIGITS flag
Due to the limiting flexibility that arduino IDEs provide, configuring compiling priorities is often not an option.

As a result, setting USED_DIGITS wouldn't get picked up by display.h. This caused array out of bounds exceptions
which further resulted in segments array to be pushed by an offset of two bytes for all member methods of the
display class, causing the first segment to no longer to function properly.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>